WEBPENTAS-backend/requirements.txt
Flask==2.0.1
SQLAlchemy==1.4.23
Flask-SQLAlchemy==2.5.1
pymysql==1.0.2
python-dotenv==0.19.0
Werkzeug==2.0.1
MySQL-Connector-Python==8.0.26 